Layout playlist details screen

+++ app/src/main/java/data/db/dao/PlaylistTracksDao.kt
@@ -13,6 +13,6 @@ interface PlaylistTracksDao {
     suspend fun add(track: PlaylistTrackEntity)
 
     @Query("SELECT * FROM playlist_tracks WHERE trackId = :id")
-    suspend fun getById(id: Int): PlaylistTrackEntity?
+    suspend fun getById(id: Long): PlaylistTrackEntity?
 
 }
